ImageVisualSearch/requirements.txt
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
faiss-cpu==1.7.4
albumentations==1.3.1
gradio==4.7.1
requests==2.31.0
beautifulsoup4==4.12.2
matplotlib==3.8.1
seaborn==0.13.0
tqdm==4.66.1
python-docx==1.1.0
pandas==2.1.3
pytest==7.4.3
